move sync logstash client to sync.go
