+++ libbeat/outputs/logstash/sync.go
